// models/MapFindLocation.ts
export type MapFindLocation = {
  distance: number | null;
  _id: string;
  id: number;
  businessName: string;
  businessCategory: string;
  businessDescription: string;
  businessWebAddress: string;
  businessPhone: string;
  businessAddress: string;
  businessLogoFileUrl: string | null;
  submitedOn: string | null;
  yourName: string | null;
  yourPositionTitle: string | null;
  yourPhone: string | null;
  yourEmail: string | null;
  showOnMap: boolean | null;
  howDidYouHearAboutUs: string | null;
  source: string | null;
  latitude: number;
  longitude: number;
  geoStatus: string;
  geoMessage: string | null;
  geoHouseNumber: string | null;
  geoZipCode: string | null;
  geoCity: string | null;
  geoStateProv: string | null;
  geoCounty: string | null;
  geoCountry: string | null;
  location: any | null;
};